Day 2 - Final Commit
This project is complete and ready for review. I added the rest of the style rules. They were mostly dealing with the margins, borders, background colors, font color and sizes, width, height, and floats as well as clears. I also added cross-browser extensions to allow the code to work properly in all major browsers (complete list of browsers is listed in the documentation of the wm_forms.css file). I added extensive documentation to the wm_forms.css file, listing each and every detail of what every piece of code does, and it's values. There is also documentation on the index.html file.